refactor(knowledge-base): migrate KB services to flat SDK boundary (#10991)
