+++ ai/daemons/services/GapInferenceEngine.mjs
@@ -1,6 +1,6 @@
 import Base                                                        from '../../../src/core/Base.mjs';
 import {Memory_Config as aiConfig, Memory_GraphService as GraphService} from '../../services.mjs';
-import KBRecorderService                                           from '../../mcp/server/knowledge-base/services/KBRecorderService.mjs';
+import KBRecorderService                                           from '../../services/knowledge-base/KBRecorderService.mjs';
 import logger                                                      from '../../mcp/server/memory-core/logger.mjs';
 
 /**
