+++ ai/services/knowledge-base/DatabaseService.mjs
@@ -1,20 +1,20 @@
-import aiConfig                  from '../config.mjs';
-import Base                      from '../../../../../src/core/Base.mjs';
+import aiConfig                  from '../../mcp/server/knowledge-base/config.mjs';
+import Base                      from '../../../src/core/Base.mjs';
 import ChromaManager             from './ChromaManager.mjs';
-import DestructiveOperationGuard from '../../shared/services/DestructiveOperationGuard.mjs';
+import DestructiveOperationGuard from '../../mcp/server/shared/services/DestructiveOperationGuard.mjs';
 import VectorService             from './VectorService.mjs';
-import ApiSource                 from '../source/ApiSource.mjs';
-import ConceptSource             from '../source/ConceptSource.mjs';
-import DiscussionSource          from '../source/DiscussionSource.mjs';
-import LearningSource            from '../source/LearningSource.mjs';
-import PullRequestSource         from '../source/PullRequestSource.mjs';
-import ReleaseNotesSource        from '../source/ReleaseNotesSource.mjs';
-import TestSource                from '../source/TestSource.mjs';
-import TicketSource              from '../source/TicketSource.mjs';
+import ApiSource                 from '../../mcp/server/knowledge-base/source/ApiSource.mjs';
+import ConceptSource             from '../../mcp/server/knowledge-base/source/ConceptSource.mjs';
+import DiscussionSource          from '../../mcp/server/knowledge-base/source/DiscussionSource.mjs';
+import LearningSource            from '../../mcp/server/knowledge-base/source/LearningSource.mjs';
+import PullRequestSource         from '../../mcp/server/knowledge-base/source/PullRequestSource.mjs';
+import ReleaseNotesSource        from '../../mcp/server/knowledge-base/source/ReleaseNotesSource.mjs';
+import TestSource                from '../../mcp/server/knowledge-base/source/TestSource.mjs';
+import TicketSource              from '../../mcp/server/knowledge-base/source/TicketSource.mjs';
 import crypto                    from 'crypto';
 import dotenv                    from 'dotenv';
 import fs                        from 'fs-extra';
-import logger                    from '../logger.mjs';
+import logger                    from '../../mcp/server/knowledge-base/logger.mjs';
 import path                      from 'path';
 import readline                  from 'readline';
