+++ ai/services/knowledge-base/QueryService.mjs
@@ -1,8 +1,8 @@
 import fs                   from 'fs-extra';
-import TextEmbeddingService from '../../memory-core/services/TextEmbeddingService.mjs';
-import mcConfig             from '../../memory-core/config.mjs';
-import aiConfig             from '../config.mjs';
-import Base                 from '../../../../../src/core/Base.mjs';
+import TextEmbeddingService from '../../mcp/server/memory-core/services/TextEmbeddingService.mjs';
+import mcConfig             from '../../mcp/server/memory-core/config.mjs';
+import aiConfig             from '../../mcp/server/knowledge-base/config.mjs';
+import Base                 from '../../../src/core/Base.mjs';
 import ChromaManager        from './ChromaManager.mjs';
 import dotenv               from 'dotenv';
 import path                 from 'path';
