+++ test/playwright/ai/neural-link.spec.mjs
@@ -6,7 +6,7 @@ import {fileURLToPath}        from 'url';
 
 const __filename  = fileURLToPath(import.meta.url);
 const __dirname   = path.dirname(__filename);
-const SERVER_PATH = path.resolve(__dirname, '../../../ai/mcp/server/neural-link/mcp-stdio.mjs');
+const SERVER_PATH = path.resolve(__dirname, '../../../ai/mcp/server/neural-link/mcp-server.mjs');
 
 test.describe('Neural Link MCP', () => {
     let client;
